XML: Allow skipping inner elements and skip over comments
If a user decides to skip over an element while parsing, we should be able to recover, so that we get the next element assuming that the user didn't skip over too much data. If the next element is a comment, we should just skip over it, instead of returning an exception for an unexpected state.

Issue: DOTNET-737