XmlTreeNode: Fix ProcessTextElement could reference null
Fix the case if processing a text element, if the handler is undefined but tree settings are defined. The code previously could call a null handler.

Additionally, introduce an `XmlContext` that is used when parsing the XML file. The context contains the information needed when reading, reducing the number of parameters that are needed between function calls.

Issue: DOTNET-737