XML: Call ProcessTextElement even if there is no text defined
The user may define a `ProcessTextElement`, which they may want to still have executed, even if the element is empty, or has only white space.

There is an API change by calling the delegate at the end. In the case that the `ProcessTextElement` is called, the position of the `XmlReader` in the `XmlNodeEventArgs` may not be what the user expects, so it must be additionally checked.

Issue: DOTNET-737