XML: Provide the user the Text on ProcessTextElement.
The user can't reliably use `e.Reader.Value` when receiving `ProcessTextElement`, because the reader may not be at the position they expect (if there was no text, then the reader is likely near the end of the element, or at the next element). Therefore, provide the value to the user that they don't need this implementation detail.

Issue: DOTNET-737