XmlTreeReader: Rework `null` constants
The constant types were intended to improve readability for multiple nulls by describing the type. A better way that also satisfies the analysers is to set to `null` direct, and reference the parameter names. This makes it easier to identify which of the more "general" method will be called.

Issue: DOTNET-492